requirements.txt
langchain==0.0.234
openai==0.27.8
chromadb==0.3.29
pypdf==3.12.2
streamlit==1.24.1
tiktoken==0.4.0
pytesseract
Pillow
python-docx
